Create MovieData Class for better of consuming the movie api
To do better use and consumption of the movie api,
i have created this class using jsonschema2pojo. 
From the first review for PopularMovies Stage 1, it
was recommended by the reviewer to use Retrofit,
instead of the AsyncTask based data downloader class.
Next commit will introduce the usage of this helper.